break out sessions.json api endpoint
